requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
DateTime==4.3
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
pytz==2019.2
requests==2.22.0
urllib3==1.25.3
Werkzeug==0.15.5
zope.interface==4.6.0
